+++ lib/searchkick/query.rb
@@ -102,11 +102,12 @@ module Searchkick
                   shared_options.merge(boost: 10 * factor, analyzer: "searchkick_search"),
                   shared_options.merge(boost: 10 * factor, analyzer: "searchkick_search2")
                 ]
-                if options[:misspellings] != false
-                  distance = (options[:misspellings].is_a?(Hash) && options[:misspellings][:distance]) || 1
+                misspellings = options.has_key?(:misspellings) ? options[:misspellings] : options[:mispellings] # why not?
+                if misspellings != false
+                  edit_distance = (misspellings.is_a?(Hash) && (misspellings[:edit_distance] || misspellings[:distance])) || 1
                   qs.concat [
-                    shared_options.merge(fuzziness: distance, max_expansions: 3, analyzer: "searchkick_search"),
-                    shared_options.merge(fuzziness: distance, max_expansions: 3, analyzer: "searchkick_search2")
+                    shared_options.merge(fuzziness: edit_distance, max_expansions: 3, analyzer: "searchkick_search"),
+                    shared_options.merge(fuzziness: edit_distance, max_expansions: 3, analyzer: "searchkick_search2")
                   ]
                 end
               elsif field.end_with?(".exact")